+++ test_test.go
@@ -105,6 +105,19 @@ var _ = Describe("type Test", func() {
 					"--- ASSERTION REPORT ---\n\n✓ pass unconditionally\n\n",
 				))
 			})
+
+			It("can be called without making an assertion", func() {
+				test.AdvanceTimeBy(
+					3*time.Second,
+					nil,
+				)
+				Expect(t.Logs).To(ContainElement(
+					"--- ADVANCING TIME BY 3s ---",
+				))
+				Expect(t.Logs).NotTo(ContainElement(
+					"--- ASSERTION REPORT ---\n\n✓ pass unconditionally\n\n",
+				))
+			})
 		})
 
 		Describe("func AdvanceTimeTo()", func() {
@@ -117,6 +130,19 @@ var _ = Describe("type Test", func() {
 					"--- ADVANCING TIME TO 2100-01-02T03:04:05Z ---",
 				))
 				Expect(t.Logs).To(ContainElement(
+					"--- ASSERTION REPORT ---\n\n✓ pass unconditionally\n\n",
+				))
+			})
+
+			It("can be called without making an assertion", func() {
+				test.AdvanceTimeTo(
+					time.Date(2100, 1, 2, 3, 4, 5, 6, time.UTC),
+					nil,
+				)
+				Expect(t.Logs).To(ContainElement(
+					"--- ADVANCING TIME TO 2100-01-02T03:04:05Z ---",
+				))
+				Expect(t.Logs).NotTo(ContainElement(
 					"--- ASSERTION REPORT ---\n\n✓ pass unconditionally\n\n",
 				))
 			})